Adding first stable version of the Java Book. This is the code for Android application which gives you an interface for a Java learning book.
This book can be downloaded as individual file at http://math.hws.edu/javanotes/

©1996--2014, David J. Eck.
This work is licensed under a Creative Commons Attribution-Noncommercial-ShareAlike 3.0 License. (This license allows you to redistribute this book in unmodified form for non-commercial purposes. It allows you to make and distribute modified versions for non-commercial purposes, as long as you include an attribution to the original author, clearly describe the modifications that you have made, and distribute the modified work under the same license as the original. Permission might be given by the author for other uses. See the license for full details.)

The most recent version of this book is always available, at no charge, for downloading and for on-line use at the Web address: http://math.hws.edu/javanotes/

Signed-off-by: Saad Shahid <saad.shahid@aim.com>